Split test functions into two more categories.
