+++ pallets/stateful-storage/src/tests/mod.rs
@@ -3,6 +3,8 @@ pub mod mock;
 mod test_common;
 
 mod apply_item_actions_tests;
+mod child_tree_tests;
 mod delete_page_tests;
+mod itemized_operations_tests;
 mod other_tests;
 mod upsert_page_tests;
